candle_api/requirements.txt
tensorflow-cpu
numpy
pandas
scikit-learn
fastapi
uvicorn
python-multipart
joblib
python-dotenv
